override hyrax methods
